Made chat floater work
